+++ src/steps/vmdr/types.ts
@@ -9,7 +9,10 @@
  * `jobState` except as mapped relationship `targetEntity` properties, and are
  * therefore not available for lookup later.
  */
-export type HostAssetTargets = { fqdn?: string; ec2InstanceArn?: string };
+export type HostAssetTargets = {
+  fqdn?: string;
+  ec2InstanceArn?: string;
+};
 
 /**
  * Maps a QWeb Host ID to `HostAssetTargets`.